refactor non-config aspects out of PerformanceConfig
