+++ .jenkins/driver_configs/baroclinic_c48_6ranks_dycore_only_serialbox.yaml
@@ -14,7 +14,7 @@ initialization:
     path: /project/s1053/fv3core_serialized_test_data/8.1.1/c48_6ranks_baroclinic/driver
     serialized_grid: grid_option
 performance_config:
-  performance_mode: false
+  collect_performance: false
 nx_tile: 48
 nz: 79
 dt_atmos: 225
