+++ driver/examples/configs/baroclinic_c12_from_serialbox.yaml
@@ -10,7 +10,7 @@ initialization_config:
   path: /test_data/8.0.0/c12_6ranks_baroclinic_dycore_microphysics/driver
   serialized_grid: True
 performance_config:
-  performance_mode: true
+  collect_performance: false
 nx_tile: 12
 nz: 79
 dt_atmos: 225
