+++ driver/examples/configs/baroclinic_c12_null_comm.yaml
@@ -8,7 +8,7 @@ stencil_config:
 initialization:
   type: baroclinic
 performance_config:
-  performance_mode: false
+  collect_performance: false
   experiment_name: c12_baroclinic
 comm_config:
   type: null_comm
