+++ driver/examples/configs/baroclinic_c12_orch_cpu.yaml
@@ -8,7 +8,7 @@ stencil_config:
 initialization:
   type: baroclinic
 performance_config:
-  performance_mode: false
+  collect_performance: false
 nx_tile: 12
 nz: 79
 dt_atmos: 225
