+++ driver/examples/configs/baroclinic_c48_6ranks_serialbox_test.yaml
@@ -14,7 +14,7 @@ initialization:
     path: notapath
     serialized_grid: true
 performance_config:
-  performance_mode: false
+  collect_performance: false
 nx_tile: 48
 nz: 79
 dt_atmos: 225
